Register and login functionality completed. A postman error caught me up - content type header  was not checked on, and it caused errors
