+++ config/middlewares.js
@@ -4,7 +4,7 @@ const jwtKey = require('../_secrets/keys').jwtKey;
 
 // quickly see what this file exports
 module.exports = {
-  authenticate,
+  authenticate, generateToken
 };
 
 // implementation details
@@ -24,4 +24,16 @@ function authenticate(req, res, next) {
       error: 'No token provided, must be set on the Authorization Header',
     });
   }
+}
+
+function generateToken(user) {
+  const payload = {
+    username: user.username,
+  };
+  const options = {
+    expiresIn: "1h",
+    jwtid: "872891"
+  };
+
+  return jwt.sign(payload, jwtKey, options);
 }
