+++ config/routes.js
@@ -2,9 +2,7 @@ const axios = require('axios');
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
 const db = require('../database/dbConfig');
-const jwtKey = require('../_secrets/keys').jwtKey;
-
-const { authenticate } = require('./middlewares');
+const { authenticate, generateToken } = require('./middlewares');
 
 module.exports = server => {
   server.post('/api/register', register);
@@ -12,56 +10,46 @@ module.exports = server => {
   server.get('/api/jokes', authenticate, getJokes);
 };
 
-function generateToken(user) {
-  const payload = {
-    username: user.username,
-  };
-  const options = {
-    expiresIn: "1h",
-    jwtid: "872891"
-  };
-
-  return jwt.sign(payload, jwtKey, options);
-}
-
 
 function register(req, res) {
-  // implement user registration
-  //hashing password
+
   const user = req.body;
-  // const hash = bcrypt.hashSync(user.password, 10);
-  // user.password = hash;
+  const hash = bcrypt.hashSync(user.password, 10);
+  user.password = hash;
 
-    db("users")
+  db('users')
     .insert(user)
-    .then(user => {
-      const token = generateToken(user);
-      res.status(201).json(token)
+    .then(function(ids) {
+      db('users')
+        .where({ id: ids[0] })
+        .first()
+        .then(user => {
+          const token = generateToken(user);
+          res.status(201).json(token);
+        });
     })
-    .catch(err => {
-      res.status(500).json({err:"Catch"})
-    })
-
-
+    .catch(function(error) {
+      res.status(500).json({ error: "Catch" });
+    });
 }
 
 function login(req, res) {
   // implement user login
-const credentials = req.body;
-db('users')
-.where({ username: credentials.username })
-.first()
-.then(function(user) {
-  if (user && bcrypt.compareSync(credentials.password, user.password)) {
-    const token = generateToken(user);
-    res.send(token);
-  } else {
-    return res.status(401).json({ error: "Invalid credentials" });
-  }
-})
-.catch(function(error) {
-  res.status(500).json({ error });
-});
+  const credentials = req.body;
+  db("users")
+    .where({ username: credentials.username })
+    .first()
+    .then(function(user) {
+      if (user && bcrypt.compareSync(credentials.password, user.password)) {
+        const token = generateToken(user);
+        res.send(token);
+      } else {
+        return res.status(401).json({ error: "Invalid credentials" });
+      }
+    })
+    .catch(function(error) {
+      res.status(500).json({ error: "catch" });
+    });
 }
 
 function getJokes(req, res) {
